VUU-18 move type in own file and better naming for item
