+++ vuu-ui/packages/vuu-shell/src/layout-management/index.ts
@@ -1,2 +1,3 @@
 export * from "./SaveLayoutPanel";
-export * from "./LayoutList"
+export * from "./LayoutList";
+export * from "./layoutTypes";
