+++ vuu-ui/packages/vuu-shell/src/layout-management/layoutTypes.ts
@@ -1,0 +1,8 @@
+export type LayoutMetadata = {
+    name: string,
+    group: string,
+    screenshot: string,
+    user: string,
+    date: string,
+    id: string
+}
